moved required/optional file obj to constants file
