+++ src/applications/ivc-champva/10-10D/components/Applicant/applicantFileUpload.js
@@ -3,7 +3,7 @@ import { titleUI } from 'platform/forms-system/src/js/web-component-patterns';
 import { applicantWording } from '../../../shared/utilities';
 import ApplicantField from '../../../shared/components/applicantLists/ApplicantField';
 import { fileUploadUi as fileUploadUI } from '../../../shared/components/fileUploads/upload';
-import { requiredFiles } from '../../config/requiredUploads';
+import { requiredFiles } from '../../config/constants';
 import { uploadWithInfoComponent } from '../Sponsor/sponsorFileUploads';
 
 // This file contains the ui/schemas for applicant file upload screens.
