+++ src/applications/ivc-champva/10-10D/components/FileUploadWrapper.jsx
@@ -1,5 +1,5 @@
 import FileFieldCustom from '../../shared/components/fileUploads/FileUpload';
-import { requiredFiles } from '../config/requiredUploads';
+import { requiredFiles } from '../config/constants';
 
 // Wrap shared fileFieldCustom so we can pass the form-specific
 // list of required uploads (for use with MissingFileOverview)
